+++ src/panel.c
@@ -284,9 +284,9 @@ void init_panel_size_and_position(Panel *panel)
 			panel->area.height = 32;
 		}
 		if (panel->fractional_width)
-			panel->area.width = (float)server.monitors[panel->monitor].width * panel->area.width / 100;
+			panel->area.width = server.monitors[panel->monitor].width * panel->area.width / 100;
 		if (panel->fractional_height)
-			panel->area.height = (float)server.monitors[panel->monitor].height * panel->area.height / 100;
+			panel->area.height = server.monitors[panel->monitor].height * panel->area.height / 100;
 		if (panel->area.width + panel->marginx > server.monitors[panel->monitor].width)
 			panel->area.width = server.monitors[panel->monitor].width - panel->marginx;
 		if (panel->area.bg->border.radius > panel->area.height / 2) {
@@ -306,12 +306,12 @@ void init_panel_size_and_position(Panel *panel)
 		}
 		int old_panel_height = panel->area.height;
 		if (panel->fractional_width)
-			panel->area.height = (float)server.monitors[panel->monitor].height * panel->area.width / 100;
+			panel->area.height = server.monitors[panel->monitor].height * panel->area.width / 100;
 		else
 			panel->area.height = panel->area.width;
 
 		if (panel->fractional_height)
-			panel->area.width = (float)server.monitors[panel->monitor].width * old_panel_height / 100;
+			panel->area.width = server.monitors[panel->monitor].width * old_panel_height / 100;
 		else
 			panel->area.width = old_panel_height;
